Refactor vim colorscheme
The vim colorscheme was mainly auto-generated and tweaked from there,
as a result the formatting was a bit awkward and the highlighting groups
were not structured clearly enough.

Now everything has a fixed structure and is formatted in a way where it
should be easy to quickly see the specified fg/bg color in cterm codes.

The highlighting groups also have been restructured and comments have
been added.